fix: c99 array designators used in C++
